Added a work-around for SATA controllers that seem to pretend to have two ATAPI drives on an emulated bus that actually only has one: disable an ATAPI device if its identify data is invalid.
git-svn-id: https://svn.aros.org/svn/aros/trunk@31258 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/common/ata.device/ata.h
@@ -295,7 +295,7 @@ struct ata_Unit
    ULONG               (*au_Write64)(struct ata_Unit *, UQUAD, ULONG, APTR, ULONG *);
    ULONG               (*au_Eject)(struct ata_Unit *);
    ULONG               (*au_DirectSCSI)(struct ata_Unit *, struct SCSICmd*);
-   ULONG               (*au_Identify)(struct ata_Unit *);
+   BYTE                (*au_Identify)(struct ata_Unit *);
 
    VOID                (*au_ins)(APTR, UWORD, ULONG);
    VOID                (*au_outs)(APTR, UWORD, ULONG);
@@ -484,8 +484,8 @@ void ata_InitBus(struct ata_Bus *);
 int atapi_SendPacket(struct ata_Unit *, APTR, APTR, LONG, BOOL*, BOOL);
 int atapi_TestUnitOK(struct ata_Unit *);
 
-ULONG atapi_Identify(struct ata_Unit*);
-ULONG ata_Identify(struct ata_Unit*);
+BYTE atapi_Identify(struct ata_Unit*);
+BYTE ata_Identify(struct ata_Unit*);
 
 ULONG atapi_DirectSCSI(struct ata_Unit*, struct SCSICmd *);
 ULONG atapi_RequestSense(struct ata_Unit* unit, UBYTE* sense, ULONG senselen);
